youtube/requirements.txt
Flask==2.3.3
google-api-python-client==2.103.0
youtube-transcript-api==0.6.1
pydantic==1.10.12
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2